Take ID generation as a value
The test suite made clear that coupling the generation of the ID
isn't really something that's great to do.
We can instead take the generator as a value,
and let the caller decide how to generate IDs.
This allows for a few things:
1. We can test sequences again.
    Before, we wouldn't be able to guess what the generated ID would be.
2. We can relax our dependencies.
    We were forcing a dependency on `purescript-uuid`, `uuid`,
    and `uuid-validate`.
    None of those are really essential to the problem we were solving.
    Instead of forcing those exact dependencies and their versions,
    we allow anything that is monadic to work.

Long and short is that we're defering what to do until later.