PS-3032 Start implementing list view
We start by defining the structure for list views.
This is a pared down implementation of what's in our specification:
https://www.notion.so/citizennet/List-View-Config-9a5c51094f9343ecb4a1615f869c29a5#5c95ed435d394f619a801c0e037e6fe2.
We're only concerning ourselves with what it will take to display
any arbitrary field.

We take a slightly different approach to how we do serialization.
In the past,
we've had the serialization layer and the business logic tightly coupled.
This works fine when we don't have any real dependents.
It doesn't work very well once we have dependents.
Any changes we make to the business logic are directly reflected in
the serialization.

This caused extra work for us most recently when we renamed some fields.
It meant that OMS had to wipe out a bunch of data in its data store.
Neverminding the fact that OMS should not be storing data in JSON blobs,
we still shouldn't be hostile to someone consuming our package.

To try and mitigate these problems in the future,
we start with a separation right now for list views.
Hopefully, when the code evolves,
it won't evolve in a way that forces us to constantly update dependents
for things they aren't really concerned about.