PS-3032 Add a halogen interpretation
We want to be able to render these views.
However, we don't want our logic to be so coupled to halogen either.
That tends to bring quite a bit of incidental complexity base on what
halogen requires.
It's already requirinb quite a lot of stuff from us now,
and we're barely using it.

To mitigate the explosion of halogen, we make another seam for rendering.
The hope is that we don't have to constantly alter the core logic
whenever we need to alter the rendering.

We also add an example to our running examples that we've got.
There's not much to see, but this is where we are.