PS-3032 Provide a way to lookup a `Field _`
We're going to need to grab the `Field _` out of a `Page _`.
That `Field _` might not exist, but we can try it anyway.

Ideally, we'd have a function where you pass it a `Key`
and you get back `Maybe ExprType`.
A good name for that might be `getValue`.
Unfortunately, `getValue` is taken.
Instead of stymieing on naming, we write this and push off to the future.